fixes for user administration
